Use relative paths for some asset loads. A bunch of casts to get rid of annoying build warnings (this will probably bite me at some point.)

+++ GCRT/box.cpp
@@ -22,7 +22,7 @@ void Box::Create()
     GenUVs(uvs);
     GenTans(tans);
 
-    numVerts = pos.size();
+    numVerts = (uint32_t)pos.size();
 
     InitVertexObjects(pos, norms, uvs, tans);
     InitModelMatrices();
